Packaging: Add new windows build script from Lawrence Rust
This is an entirely new build script that will compile and build
installation packages for mythtv and mythplugins natively on a windows
machine and cross-compile from Linux. There is additional support for
building natively for Linux and cross-compiling for OS X (untested).

This version is based on Lawrence's git master:-

http://www.softsystem.co.uk/cgi-

with the following changes:-

- the build defaults to master (rather than fixes/0.24)
- patches will NOT be applied to a master build (only stable releases
such as fixes/0.24)
- building myththemes is disabled by default and enabled with the -T
option (this is to encourage use of the theme chooser/downloader and
reduce the size of the package for distribution)

At some point in the future, the ability to patch the mythtv source will
be removed entirely.