Bump 0.24.1 and 0.25 ebuilds
This commit bumps the ebuilds to the latest versions as of current.
This commit fixes the ebuild generator to properly create new mythvideo
ebuilds for 0.24.1.  This also adds proper log rotation for the frontend
and backend following the new logging format, and the inclusion of a
script to remove old copies of the logs from previous instances of those
applications.