Correct version version string generation
Github changed their tarball download mechanism from providing the name
of the file directly in the URL, to providing the name in the headers.
This filename is used to specify the version string that the build
scripts otherwise expect to pull out of git, and the change resulted in
an invalid, chunk of SHA1 to be used instead. This corrects the
extraction of the version string by pulling it out of the new location
in the headers.