Update mac build system to work on 10.7 Lion.
Add support for all version of XCode (XCode 3 and 4)

Fully rework the build script to automatically determine the best compiler to use and the SDK that is compatible with most versions of OS X (e.g. the oldest)

The core change is upgrading to Qt 4.8. This means that supports for 10.4 does no longer exists. Having said that, I'm pretty convinced that it hasn't worked on 10.4 in a long time...

Add a README.txt to document basic usage.

Ultimately, this should be all redone so it uses public Qt builds instead of spending countless of hours trying to compile Qt ourselves.

The old build package using Qt 4.7.4 and that could still support 10.4 has been renamed osx-packager-10.4.pl.