New mac packager using external Qt libraries
This is a major rewrite of the mac packager. It now uses external Qt libraries as downloaded from the Qt web site.

It adds significant improvement:
- 64 bits application support (both PPC and Intel)
- Universal 32/64 bits support

So far tested with XCode 4.3 on OSX 10.7 and XCode 3.2 on OS X 10.6

Application built on 10.6 will run on 10.5, 10.6 and 10.7
Application built on 10.7 will run on 10.6 and 10.7. There is an easy fix to the 10.6 SDK that can make the applications compatible on 10.5. This will be documented later.

For the time being, the builder only works with very recent master/0.25 and several fixes were added to the code to properly support LLVM compiler.
0.24 support may be added later