Fix wrong dependency to one of mythtv[php] or mythtv-bindings[php].
The other format was merely the first one that appeared to work with
'emerge -vp', after stumbling upon it from trial and error. To be
honest, I've never actually installed Apache on a Gentoo system before,
so the limited amount I've been changing on the MythWeb ebuilds is
finding something emerge doesn't complain about.

Fixes #10785

Signed-off-by: Raymond Wagner <rwagner@mythtv.org>