Big ebuild bump.
This includes a long overdue update of the ebuilds for 0.25.1 and 0.26.
This discontinues the older set of 0.24 ebuilds, as well as the
MythVideo plugin, which is included as part of core in 0.25 and beyond.

The ebuild generator 'mythtv-buildebuild.py' is no longer viable due to
changes in the Github API, so it has been removed, to be replaced
shortly with an updated version.

This updates to a versio of 0.26_pre that fixes #10507 and the fix is
included as a patch for 0.25.1.

This resets the MythNews ebuild back to the 0.24 version, as it had
accidentally been overwritten by the ebuild for MythGallery.  It also
adds MythBrowser as a prerequisite.