deb: let upstart handle the backend process correctly
When shutting down the signal got send to the shell that was used in
starting the backend. As the shell immediately ends the kill after
kill timeout part of the shutdown process never happened.

This bug was exposed by the long shutdown times in master.
Inspiration for the fix to the upstart job by bill6502 on irc.
Fix for the upstart job by Thomas Mashos.