Discover the android tools version at build time.
The build scripts will now look into the tools directory to see how
many versions of the tools are present.  If there is only one version
of tools, then the build scripts will use it.  If there is more than
one version of the tools, the build script will fall back to the
hard-coded version.