deb: Fix debian/changelog parsing
[ $EPOCH ':' ] [ '0.' ] $MAJOR '.' $MINOR ( '+fixes' | '~master' ) [ '.' $YEAR $MONTH $DAY '.' $hash '-0ubuntu' $COUNTER ]

The old code tries to parse the package version, which sometimes(!)
contains a time-stamp, which then is used to find the additional
changes since then for the Debian packaging.
This currently is broken and always(!) appends all git commits to the
changelog.

Instead of parsing the version string, use the date from the last
changelog entry and use that.