Fix startup error in Android with Qt 5.15.x.
This is a known bug(*) with Qt 5.15 that causes libraries to not be
found when android:extractNativeLibs is left to default to false.  It
is not clear why some other developers are not affected by this.
Maybe something in their environment makes it default to true.

(*)See https://bugreports.qt.io/browse/QTBUG-61072 for details.