Android: Add Qt mysql driver patch
The MariaDB-connector-c version 3.2 and MariaDB server version 10.6
cooperate to avoid re-transferring the query metadata, so the fact that
we were modifying it was causing it to improperly decode the DATETIME
data types into string, as we had asked. We ended up with a 7-byte
string that was actually the date binary-encoded.

Fixes: QTBUG-95639
Ref: https://invent.kde.org/qt/qt/qtbase/-/commit/49a3ddfa8d1ee992746475adf09ebbdefbecc60b