Move Windows build scripts from MythTV repo to packaging
The Windows build scripts are currently in both the main repo and in the
packaging repo. The other platforms have already had their build scripts
moved to the packaging repo.

1) Move the Windows build scritps from MythTV repo to packaging repo for
consistency with other platforms.

2) Minor refactor of the Win32 directotry structure to allow merging of
the two repos.

A future commit will be raised to remove the platform directory from the
main MythTV repo once this commit is accepted.